+++ example/android/app/src/main/kotlin/com/appplayer/example/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.appplayer.example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()